moved StopWatch.tsx and StopWatchButton.tsx into a components folder
