+++ src/App.tsx
@@ -1,10 +1,8 @@
-import React from 'react'
-import './App.css'
-import StopWatch from './StopWatch'
+import React from "react";
+import "./App.css";
+import StopWatch from "./components/StopWatch";
 
 //default function
 export default function App() {
-    return(
-        <StopWatch />
-    )
+  return <StopWatch />;
 }
