+++ src/StopWatchButton.test.tsx
@@ -1,7 +1,6 @@
 import { render, fireEvent, screen } from "@testing-library/react";
-import StopWatchButton from "./StopWatchButton";
+import StopWatchButton from "./components/StopWatchButton";
 import React from "react";
-
 
 //tests the StopWatchButton functions
 describe("testing StopWatchButton functionality", () => {
@@ -94,21 +93,32 @@ describe("testing StopWatchButton functionality", () => {
     expect(setLapsMock).toHaveBeenCalledWith([]);
   });
 
-  test('Stopwatch starts  correctly', () => {
-    const { getByText } = render(<StopWatchButton time={0} setTime={() => {}}  laps={[]} setLaps={() => {}} />);
-    const startButton = getByText('Start');
-      fireEvent.click(startButton);
+  test("Stopwatch starts  correctly", () => {
+    const { getByText } = render(
+      <StopWatchButton
+        time={0}
+        setTime={() => {}}
+        laps={[]}
+        setLaps={() => {}}
+      />
+    );
+    const startButton = getByText("Start");
+    fireEvent.click(startButton);
     expect(startButton).toBeDefined();
-  
   });
-  
-  test('Stopwatch resets correctly', () => {
-    const { getByText } = render(<StopWatchButton time={10} setTime={() => {}}  laps={[]} setLaps={() => {}} />);
-    const resetButton = getByText('Reset');
+
+  test("Stopwatch resets correctly", () => {
+    const { getByText } = render(
+      <StopWatchButton
+        time={10}
+        setTime={() => {}}
+        laps={[]}
+        setLaps={() => {}}
+      />
+    );
+    const resetButton = getByText("Reset");
 
     fireEvent.click(resetButton);
-    expect(getByText('Start')).toBeDefined();
+    expect(getByText("Start")).toBeDefined();
   });
-  
- 
 });
